correcciones

+++ EjerciciosT9/src/ejerciciost9/TraduccionImplESP.java
@@ -4,7 +4,7 @@ import java.util.Scanner;
 
 public class TraduccionImplESP implements Traduccion {
     private  double distancia;
-    private int tiempo;
+    private double tiempo;
     private double velocidad;
     
     @Override
@@ -25,7 +25,7 @@ public class TraduccionImplESP implements Traduccion {
     public void inicioRespuesta() {
         final int CST = 60;
         velocidad = distancia/(tiempo/CST);
-        velocidad = Math.round(velocidad*100)/100;
+        velocidad = Math.round(velocidad*100)/100.;
     }
 
     @Override
